New functionality
Added way to have it look in a "statements" folder alongside the script and list out the ones it finds in newest to oldest so user can choose what to load

Added checking of dupes and ordering the output by date before writing to file.

Dupe check allows user to input y/n to keep the input or not.